moved predeploys to its own file. moved functions relating to proofs to a new file. moved withdraw function to a new file and renamed it. renamed a few other functions.
