+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './proof'
+export * from './l2-to-L1-message-relaying'
 export * from './string-format'
 export * from './rlp'
