+++ src/l2-to-L1-message-relaying.ts
@@ -1,89 +1,21 @@
-import { ethers, BigNumber } from 'ethers'
-import * as rlp from './rlp'
-import { toHexString, toRpcHexString, remove0x } from './string-format'
+import { ethers } from 'ethers'
 import { setTxOptionsForL2, formatNVMTx, formatNVMReceipt } from './l2-tx-formatting'
+import { CrossDomainMessageProof, GetTransactionProof } from './tx-proof'
+import { predeploys } from './predeploys'
 
 import L1CrossDomainMessengerMetadata from './contract-metadata/iNVM_L1CrossDomainMessenger.json'
 import L2CrossDomainMessengerMetadata from './contract-metadata/NVM_L2CrossDomainMessenger.json'
-import L2StandardBridgeMetadata from './contract-metadata/NVM_L2StandardBridge.json'
-import type = Mocha.utils.type
-
-interface StateTrieProof {
-  accountProof: string
-  storageProof: string
-}
 
 interface CrossDomainMessagePair {
   message: CrossDomainMessage
   proof: CrossDomainMessageProof
 }
 
-interface CrossDomainMessage {
+export interface CrossDomainMessage {
   target: string
   sender: string
   message: string
   messageNonce: number
-}
-
-interface CrossDomainMessageProof {
-  stateRoot: string
-  stateTrieWitness: string
-  storageTrieWitness: string
-}
-
-const predeploys = {
-  NVM_L2ToL1MessagePasser: '0x4200000000000000000000000000000000000000',
-  NVM_L1MessageSender: '0x4200000000000000000000000000000000000001',
-  NVM_DeployerWhitelist: '0x4200000000000000000000000000000000000002',
-  NVM_ECDSAContractAccount: '0x4200000000000000000000000000000000000003',
-  NVM_SequencerEntrypoint: '0x4200000000000000000000000000000000000005',
-  NVM_ETH: '0x4200000000000000000000000000000000000006',
-  NVM_L2CrossDomainMessenger: '0x4200000000000000000000000000000000000007',
-  Lib_AddressManager: '0x4200000000000000000000000000000000000008',
-  NVM_ProxyEOA: '0x4200000000000000000000000000000000000009',
-  NVM_ExecutionManagerWrapper: '0x420000000000000000000000000000000000000B',
-  NVM_GasPriceOracle: '0x420000000000000000000000000000000000000F',
-  NVM_SequencerFeeVault: '0x4200000000000000000000000000000000000011',
-  NVM_L2StandardBridge: '0x4200000000000000000000000000000000000010',
-  ERC1820Registry: '0x1820a4B7618BdE71Dce8cdc73aAB6C95905faD24',
-}
-
-/**
- * Generates a Merkle-Patricia trie proof for a given account and storage slot.
- *
- * @param l2RpcProvider L2 RPC provider.
- * @param blockNumber Block number to generate the proof at.
- * @param address Address to generate the proof for.
- * @param slot Storage slot to generate the proof for.
- * @returns Account proof and storage proof.
- */
-const getStateTrieProof = async (
-  l2RpcProvider: ethers.providers.JsonRpcProvider,
-  blockNumber: number,
-  address: string,
-  slot: string
-): Promise<StateTrieProof> => {
-  const proof = await l2RpcProvider.send('eth_getProof', [address, [slot], toRpcHexString(blockNumber)])
-
-  return {
-    accountProof: toHexString(rlp.encode(proof.accountProof)),
-    storageProof: toHexString(rlp.encode(proof.storageProof[0].proof)),
-  }
-}
-
-/**
- * Encodes a cross domain message.
- *
- * @param message Message to encode.
- * @returns Encoded message.
- */
-const encodeCrossDomainMessage = (message: CrossDomainMessage): string => {
-  return new ethers.utils.Interface(L2CrossDomainMessengerMetadata.abi).encodeFunctionData('relayMessage', [
-    message.target,
-    message.sender,
-    message.message,
-    message.messageNonce,
-  ])
 }
 
 /**
@@ -154,51 +86,6 @@ export const getL2ToL1MessagesByBlock = async (
 }
 
 /**
- * Generate proofs for transactions
- *
- * @param message A CrossDomainMessage from the transaction
- * @param receipt transaction receipt
- * @param l2CrossDomainMessengerAddress
- * @param l2RpcProvider
- * @constructor
- */
-const GetTransactionProof = async (
-  message: CrossDomainMessage,
-  receipt: ethers.providers.TransactionReceipt,
-  l2CrossDomainMessengerAddress: string,
-  l2RpcProvider: ethers.providers.JsonRpcProvider
-): Promise<CrossDomainMessageProof> => {
-  // We need to calculate the specific storage slot that demonstrates that this message was
-  // actually included in the L2 chain. The following calculation is based on the fact that
-  // messages are stored in the following mapping on L2:
-  // https://github.com/ethereum-optimism/optimism/blob/c84d3450225306abbb39b4e7d6d82424341df2be/packages/contracts/contracts/optimistic-ethereum/OVM/predeploys/OVM_L2ToL1MessagePasser.sol#L23
-  // You can read more about how Solidity storage slots are computed for mappings here:
-  // https://docs.soliditylang.org/en/v0.8.4/internals/layout_in_storage.html#mappings-and-dynamic-arrays
-  const messageSlot = ethers.utils.keccak256(
-    ethers.utils.keccak256(encodeCrossDomainMessage(message) + remove0x(l2CrossDomainMessengerAddress)) +
-      '00'.repeat(32)
-  )
-
-  // We need a Merkle trie proof for the given storage slot. This allows us to prove to L1 that
-  // the message was actually sent on L2.
-  const stateTrieProof = await getStateTrieProof(
-    l2RpcProvider,
-    receipt.blockNumber,
-    predeploys.NVM_L2ToL1MessagePasser,
-    messageSlot
-  )
-
-  // We now have enough information to create the message proof.
-  const proof: CrossDomainMessageProof = {
-    stateRoot: receipt.root!,
-    stateTrieWitness: stateTrieProof.accountProof,
-    storageTrieWitness: stateTrieProof.storageProof,
-  }
-
-  return proof
-}
-
-/**
  * Finds all L2 => L1 messages sent in a given L2 transaction and generates proofs for each of
  * those messages.
  *
@@ -230,11 +117,31 @@ export const getMessagesAndProofsForL2Transaction = async (
 
   const messagePairs: CrossDomainMessagePair[] = []
   for (const message of messages) {
-    const proof = await GetTransactionProof(message, receipt, l2CrossDomainMessengerAddress, l2RpcProvider)
+    const proof = await GetTransactionProof(
+      encodeCrossDomainMessage(message),
+      receipt,
+      l2CrossDomainMessengerAddress,
+      l2RpcProvider
+    )
     messagePairs.push({ message, proof })
   }
 
   return messagePairs
+}
+
+/**
+ * Encodes a cross domain message.
+ *
+ * @param message Message to encode.
+ * @returns Encoded message.
+ */
+const encodeCrossDomainMessage = (message: CrossDomainMessage): string => {
+  return new ethers.utils.Interface(L2CrossDomainMessengerMetadata.abi).encodeFunctionData('relayMessage', [
+    message.target,
+    message.sender,
+    message.message,
+    message.messageNonce,
+  ])
 }
 
 /**
@@ -251,31 +158,6 @@ export const sleep = async (ms: number): Promise<void> => {
 }
 
 /**
- * Initiate withdrawals.
- *
- * @param l2TokenAddress L2 address of the to be withdrawn token.
- * @param withdrawAmount The amount to withdraw.
- * @param l2Provider L2 provider.
- * @param signer L2 transaction signer.
- * @returns Returns the transaction response containing metadata for the withdrawal transaction.
- */
-export const withdraw = async (
-  l2TokenAddress: string,
-  withdrawAmount: BigNumber,
-  l2Provider: ethers.providers.JsonRpcProvider,
-  signer: ethers.Signer
-): Promise<ethers.providers.TransactionResponse> => {
-  const L2StandardBridgeInterface = new ethers.utils.Interface(L2StandardBridgeMetadata.abi)
-  const contract = new ethers.Contract(predeploys.NVM_L2StandardBridge, L2StandardBridgeInterface, l2Provider)
-  const signerWithProvider = signer.connect(l2Provider)
-  const transactionResponse = await contract
-    .connect(signerWithProvider)
-    .withdraw(l2TokenAddress, withdrawAmount, 0, '0x')
-
-  return transactionResponse
-}
-
-/**
  * Relays all L2 => L1 messages found in a given L2 transaction.
  *
  * @param l2TransactionHash L2 transaction hash to find the messages in.
@@ -285,7 +167,7 @@ export const withdraw = async (
  * @param l1Signer L1 transaction signer.
  * @param maxRetries maximum retries for relaying messages.
  */
-export const relayXDomainMessages = async (
+export const relayL2ToL1Messages = async (
   l2TransactionHash: string,
   l1CrossDomainMessengerAddress: string,
   l1RpcProvider: ethers.providers.JsonRpcProvider,
