+++ src/predeploys.ts
@@ -1,0 +1,16 @@
+export const predeploys = {
+  NVM_L2ToL1MessagePasser: '0x4200000000000000000000000000000000000000',
+  NVM_L1MessageSender: '0x4200000000000000000000000000000000000001',
+  NVM_DeployerWhitelist: '0x4200000000000000000000000000000000000002',
+  NVM_ECDSAContractAccount: '0x4200000000000000000000000000000000000003',
+  NVM_SequencerEntrypoint: '0x4200000000000000000000000000000000000005',
+  NVM_ETH: '0x4200000000000000000000000000000000000006',
+  NVM_L2CrossDomainMessenger: '0x4200000000000000000000000000000000000007',
+  Lib_AddressManager: '0x4200000000000000000000000000000000000008',
+  NVM_ProxyEOA: '0x4200000000000000000000000000000000000009',
+  NVM_ExecutionManagerWrapper: '0x420000000000000000000000000000000000000B',
+  NVM_GasPriceOracle: '0x420000000000000000000000000000000000000F',
+  NVM_SequencerFeeVault: '0x4200000000000000000000000000000000000011',
+  NVM_L2StandardBridge: '0x4200000000000000000000000000000000000010',
+  ERC1820Registry: '0x1820a4B7618BdE71Dce8cdc73aAB6C95905faD24',
+}
